Use a list of pools per prefix for data
Each data prefix should point to a list of pools so later it will
allow to specify directly into which pool an inode might go.

This required many changes in the methods that deal with the pools
as well as the deletion of the method dataPoolFromPrefix that no
longer makes sense.