Make permissions inaccessible from XAttrs calls
Instead of having a sys. prefix, which is accessible by root,
permissions now have a priv. prefix because they should not be
accessible from the XAttrs related API. The main reason for this is
that now that files no longer have an inode necessarily at creation
time, the permissions will lie in the link XAttr corresponding to
the file, making it inconsistent with the directories' permissions
XAttr.