Implement the file rename operation
This replaces the file moving logic left after changing the
methods' name to "rename".

The reason for the change is that having a move operation might
imply moving a file to a different directory or renaming it.
This should be distinct operations and the move one should be
implemented by the user, not the library.