Stat objects according to their path's appearance
By default it was always trying to stat objects as directories and
only statting them as files after thist first approach failed.

This patch changes that and stats the objects according to what their
path looks like -- if it has an ending /, try to stat as a directory
first, otherwise try to stat as a file. This will improve performance
for cases where the paths are correctly representing their objects,
which always happen when e.g. listing and statting a directory's
contents.