Create an class to manage asynchronous operations
This new class RadosFsAsyncOperation can be used to track a list
of asynchronous operations and gives information such as the return
value, whether the operation is still running or is already finished
and can also block until it is.

Hopefully this will be passed on to the user when we offer the API to
track asynchronous operations.