Fix problem in the use count of RadosFsIO
The RadosFsIO is only immediately removed if it has just one single user
but this check was outdated as it was testing for a use count of 1 when
the RadosFsPriv's map now also holds a reference to each of its elements
too.

These changes introduce and use a static method in RadosFsIO to check if
the object has only one client.