test: Fix sync issues in FileOpsMultipleClients
Sometimes the operations (called from two threads) in this test
were called before the file lock (from a previous operation)
had become idle and removed. This would make those operations
wait for the lock until it was removed which resulted in one them
not happening in the order we wanted (e.g. truncate would happen
after a remove when the test is designed in the opposite way).