Use the FileInlineBuffer in the FileIO
If the FileInlineBuffer is set in a FileIO, then it will use it in
its operations and it will only create the inode object if the
contents to be written are bigger than the FileInlineBuffer's size.
Due to these changes, the way the mtime is set on an inode has to
be set from the FileIO object instead of FileInode.

Also adds methods for setting and getting the FileInlineBuffer.