Override certain parallel ops's return values
When reading a file in parallel, there can be certain operations
that will fail because the file's chunks do not exist. However,
depending on the operation, a nonexistent file chunk does not mean
that an error code should be returned to the user. E.g. when reading
a file that has been truncated to a value beyond the existent data; in
this example, the user does not need to know details about what actually
exists or not, only that the value stated in the file's size should be
readable.