Replace TM time from dirs by TM id
The Transversal Modification Id is similar to the TM time but uses
a UUID instead of a time stamp. The reason for this change is that
in a distributed system, the use of times for the purpose of checking
if there were changes (and followng the chain of changes) may not be
the best choice.