Improve performance when writing files asynchronously
FileIO instancea are shared among different File instances but
if one File instance is created and destroyed, the FileIO will be
promptly destroyed too, which means that it will call 'sync' on any
of its asynchronous operations. These changes only destroy the FileIO
instances if all of its asynchronous operations are finished which
improves performance in some use cases (because the execution does not
have to wait for a file's operations to be finished just before a file
instance was destroyed).