test: Run the 2nd client's actions after a delay
The multiple clients tests, sometimes the 2nd client's operation
is running before the first one effectively does something.
This is a problem before it invalidates the test's purpose so we're
delaying the notification for the 2nd client to run in order to
give time for the 1st client's operation to really start.

This is not an optimal solution and it should be changed in the
future to e.g. checking if there's a lock on the inode object that
is being modified.