Create a unit-test Binary to Run Unit Tests
Created a separate binary to run unit tests. These tests do not require a ceph
configuration file, and do not connect to the network in any way. They are meant
to test individual functions.

Since unit tests are built with different flags, two different libraries
are now built from the radosfs sources (radosfs and radosfs-ut.) The unit-tests
binary links to radosfs-ut, which is never installed.

unit-tests.cc currently have one unit test, which is to check fileSizeToHex()
for a potential buffer overflow that occurs when compiled with "-O
-Wp,-D_FORTIFY_SOURCE=2", which is the RHEL/CentOS default.

This also fixes the potential buffer overflow by increasing the target char
array to accommodate the trailing NULL that sprintf writes.